// src/app/models/institucion.ts
export class Institucion{
    id?: number;
    codUe: string;  
    nombreUe: string;
    numResolucionAdm1: number;
    fechaResolucionAdm1: string;
    codDistritoEducativo: string;
    nombreDistritoEducativo: string;
    numResolucionAdm2: number;
    fechaResolucionAdm2: string;
    codEdificioEscolar: string;
    nuevoCodEdificioEscolar: string;
    departamento: string;
    provincia: string;
    seccionMunicipal: string;
    canton: string;
    ciudad: string;
    zona: string;
    direccion: string;
    estatal: boolean;
    convenio: boolean;
    nombreInstitucion: string;
    comunitaria: boolean;
    privada: boolean;
    edFormal: boolean;
    edInicial : boolean;
    edPrimaria: boolean;
    edSecundaria: boolean; 
    edAlternativa: boolean;
    edAdultos: boolean;
    edEspecial: boolean;
    edPermanente: boolean;
    bachillerato: boolean;
    humanistico: boolean;
    tecnico: boolean;
    siglas: number;
    latitud: number;
    longitud: number;
    correo: string;
    telefono1: string;
    telefono2: string;
    celular1: string;
    celular2: string;
    web: string;
    descripcion: string;
    lugarRecepcion: string;
    fechaRecepcion: string;
    estadoSql : number;
}